client: rename some stuff
