+++ modules/client/src/Connext.ts
@@ -16,7 +16,7 @@ import SyncController from './controllers/SyncController'
 import ThreadsController from './controllers/ThreadsController'
 import WithdrawalController from './controllers/WithdrawalController'
 import {  HubAPIClient, IHubAPIClient } from './Hub'
-import { default as Logger } from './lib/Logger'
+import { Logger } from './lib/logger'
 import { Networking } from './lib/networking'
 import { isFunction, timeoutPromise } from './lib/utils'
 import * as actions from './state/actions'
