+++ modules/client/src/controllers/DepositController.test.ts
@@ -1,7 +1,7 @@
-import { assert, } from '../testing/index';
-import { MockConnextInternal, patch } from '../testing/mocks';
+import { assert } from '../testing/index'
+import { MockConnextInternal, patch } from '../testing/mocks'
 // @ts-ignore
-global.fetch = require('node-fetch-polyfill');
+global.fetch = require('node-fetch-polyfill')
 
 describe('DepositController: unit tests', () => {
   let connext: MockConnextInternal
