+++ modules/client/src/controllers/ExchangeController.test.ts
@@ -1,5 +1,5 @@
-import { mkAddress } from '../testing';
-import { MockStore, MockConnextInternal } from '../testing/mocks';
+import { mkAddress } from '../testing'
+import { MockConnextInternal, MockStore } from '../testing/mocks'
 // @ts-ignore
 global.fetch = require('node-fetch-polyfill');
