+++ modules/client/src/controllers/ExchangeController.ts
@@ -1,11 +1,11 @@
-import * as eth from 'ethers'
-import { AbstractController } from './AbstractController'
-import { ConnextInternal } from '../Connext';
-import { toBN, toWei } from '../lib/bn';
-import { getExchangeRates, getTxCount } from '../state/getters'
-import { Poller } from '../lib/poller/Poller';
+import { ethers as eth } from 'ethers'
+import { ConnextInternal } from '../Connext'
+import { toBN, toWei } from '../lib/bn'
+import { Poller } from '../lib/poller'
 import * as actions from '../state/actions'
+import { getExchangeRates, getTxCount } from '../state/getters'
 import { ConnextStore } from '../state/store'
+import { AbstractController } from './AbstractController'
 
 const ONE_MINUTE = 1000 * 60
