+++ modules/client/src/controllers/RedeemController.ts
@@ -1,12 +1,12 @@
-import { ethers as eth } from 'ethers';
-import { AbstractController } from "./AbstractController";
-import { Payment } from "../types";
+import { ethers as eth } from 'ethers'
+import { Payment } from '../types'
+import { AbstractController } from './AbstractController'
 
 export class RedeemController extends AbstractController {
   public redeem = async (secret: string): Promise<{ purchaseId: string, amount: Payment }> => {
     // check that the secret was generated as a hex
     if (!eth.utils.isHexString(secret)) {
-      throw new Error(`The secret provided is not a hex string. Was it generated using the 'generateSecret' method of connext? Secret: ${secret}`)
+      throw new Error(`The secret provided is not a hex string. Secret: ${secret}`)
     }
 
     const state = this.getState()
