+++ modules/client/src/lib/logger.ts
@@ -1,4 +1,5 @@
-export default interface Logger {
+export interface Logger {
   source: string
   logToApi(key: string, data: any): Promise<void>
 }
+export default Logger
