--- modules/client/src/lib/naming.ts
@@ -1,3 +1,0 @@
-export function capitalize(str: string): string {
-  return str.substring(0, 1).toUpperCase() + str.substring(1)
-}
