+++ modules/client/src/lib/poller.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 
-import { sleep } from '../utils'
-import { Poller } from './Poller'
+import { sleep } from './utils'
+import { Poller } from './poller'
 
 describe('Poller', () => {
   const interval: number = 10
