+++ modules/client/src/types.ts
@@ -1,6 +1,6 @@
 import { isArray, isNullOrUndefined } from 'util'
 import { BN, isBN, toBN } from './lib/bn'
-import { CurrencyType } from './lib/currency/Currency'
+import { CurrencyType } from './lib/currency'
 
 ////////////////////////////////////////
 // Export useful types defined in other modules
@@ -23,8 +23,8 @@ export {
 } from 'ethers/utils'
 
 export { ChannelManager } from './contract/ChannelManager'
-export { CurrencyType, default as Currency } from './lib/currency/Currency'
-export { default as CurrencyConvertable } from './lib/currency/CurrencyConvertable'
+export { CurrencyType, default as Currency } from './lib/currency'
+export { CurrencyConvertable } from './lib/currencyConvertable'
 
 ////////////////////////////////////////
 // Common types eg Exchange Rates
